+++ src/helper_functions.cpp
@@ -25,10 +25,10 @@ Clip commandClip(vector<string> commands)
 {
     assert(commands.size() == 7);
     FollowAction follow_actions[ACTIONS];
-    for(int i=0; i<ACTIONS; i++)
+    for(unsigned int i=0; i<ACTIONS; i++)
     {
         follow_actions[i].actionName = STRING_TO_ACTION_CONVERTER.at(commands[i+FOLLOW_ACTION_NAME]);
-        follow_actions[i].chance = stoi(commands[i+FOLLOW_ACTION_CHANCE]);
+        follow_actions[i].chance = stof(commands[i+FOLLOW_ACTION_CHANCE]);
     }
 
     return Clip(commands[CLIP_NAME], stoi(commands[CLIP_TICKS]), follow_actions);
@@ -50,7 +50,7 @@ vector<pair<float, float>> commandTicks(vector<string> commands)
         else if((commands.size()> TICKS_WITHOUT_PAIR) && (commands.size() < 2*stoi(commands[TICKS_TICKS])))
         {
             //Add the existing values
-            for(int i=TICKS_START_PAIR; i<commands.size(); i= i+2)
+            for(unsigned int i=TICKS_START_PAIR; i<commands.size(); i= i+2)
             {
                 ticks_pair.push_back(pair<float, float>(stof(commands[i]), stof(commands[i+1])));
             }
